+++ apps/database-new/app/[threadId]/layout.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { Chat } from '@/components/Chat/Chat'
+import { Chat } from '@/app/[threadId]/Chat'
 import SaveSchemaDropdown from '@/components/Header/SaveSchemaDropdown'
 import ToggleCodeEditorButton from '@/components/Header/ToggleCodeEditorButton'
 
@@ -7,7 +7,7 @@ export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <div className="flex flex-col-reverse items-between xl:flex-row xl:items-center xl:justify-between bg-alternative h-full">
       <Chat />
-      <div className="block xl:hidden flex items-center gap-x-2 justify-end border-t py-2 px-2 bg-background">
+      <div className="xl:hidden flex items-center gap-x-2 justify-end border-t py-2 px-2 bg-background">
         <ToggleCodeEditorButton />
         <SaveSchemaDropdown />
       </div>
